THRIFT-1740 Make C++ library build on OS X and iOS Patch: Ben Craig minor modification by Roger Meier
git-svn-id: https://svn.apache.org/repos/asf/thrift/trunk@1410127 13f79535-47bb-0310-9956-ffa450edef68

+++ lib/cpp/src/thrift/concurrency/Util.h
@@ -67,8 +67,8 @@ class Util {
   }
 
   static void toTimeval(struct timeval& result, int64_t value) {
-    result.tv_sec = (uint32_t)(value / MS_PER_S); // ms to s
-    result.tv_usec = (value % MS_PER_S) * US_PER_MS; // ms to us
+    result.tv_sec  = static_cast<uint32_t>(value / MS_PER_S); // ms to s
+    result.tv_usec = static_cast<uint32_t>((value % MS_PER_S) * US_PER_MS); // ms to us
   }
 
   static void toTicks(int64_t& result, int64_t secs, int64_t oldTicks,
